api_server/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
sqlalchemy==1.4.32
mysql-connector-python==8.1.0
python_dotenv==1.0.0
